Record model to InContextEditing
